Put StreamingE2E class into a new module
