+++ espnet/nets/pytorch_backend/e2e_asr.py
@@ -446,170 +446,3 @@ class E2E(ASRInterface, torch.nn.Module):
             np.array(x, dtype=np.float32)))
         h.contiguous()
         return h, ilen
-
-
-# TODO(pzelasko): Currently allows half-streaming only; needs streaming attention decoder implementation
-class WindowStreamingE2E(object):
-    """Convenience wrapper over E2E class for streaming recognitions.
-
-    Not recommended for GPUs.
-    """
-
-    def __init__(self, e2e, recog_args, rnnlm=None):
-        """WindowStreamingE2E constructor.
-
-        :param E2E e2e: E2E ASR object
-        :param recog_args: arguments for "recognize" method of E2E
-        """
-        self._e2e = e2e
-        self._recog_args = recog_args
-        self._char_list = e2e.char_list
-        self._rnnlm = rnnlm
-
-        self._e2e.eval()
-
-        self._offset = 0
-        self._previous_encoder_recurrent_state = None
-        self._encoder_states = []
-        self._ctc_posteriors = []
-        self._last_recognition = None
-
-        assert self._recog_args.ctc_weight > 0.0, \
-            "WindowStreamingE2E works only with combined CTC and attention decoders."
-
-    def accept_input(self, x):
-        """Call this method each time a new batch of input is available."""
-
-        h, ilen = self._e2e.subsample_frames(x)
-
-        # Streaming encoder
-        h, _, self._previous_encoder_recurrent_state = self._e2e.enc(
-            h.unsqueeze(0),
-            ilen,
-            self._previous_encoder_recurrent_state
-        )
-        self._encoder_states.append(h.squeeze(0))
-
-        # CTC posteriors for the incoming audio
-        self._ctc_posteriors.append(self._e2e.ctc.log_softmax(h).squeeze(0))
-
-    def _input_window_for_decoder(self, use_all=False):
-        if use_all:
-            return torch.cat(self._encoder_states, dim=0), torch.cat(self._ctc_posteriors, dim=0)
-
-        def select_unprocessed_windows(window_tensors):
-            last_offset = self._offset
-            offset_traversed = 0
-            selected_windows = []
-            for es in window_tensors:
-                if offset_traversed > last_offset:
-                    selected_windows.append(es)
-                    continue
-                offset_traversed += es.size(1)
-            return torch.cat(selected_windows, dim=0)
-
-        return (
-            select_unprocessed_windows(self._encoder_states),
-            select_unprocessed_windows(self._ctc_posteriors)
-        )
-
-    def decode_with_attention_offline(self):
-        """Run the attention decoder offline.
-
-        Works even if the previous layers (encoder and CTC decoder) were being run in the online mode.
-        This method should be run after all the audio has been consumed.
-        This is used mostly to compare the results between offline and online implementation of the previous layers.
-        """
-        h, lpz = self._input_window_for_decoder(use_all=True)
-
-        return self._e2e.dec.recognize_beam(h, lpz, self._recog_args, self._char_list, self._rnnlm)
-
-
-class SegmentStreamingE2E(object):
-    def __init__(self, e2e, recog_args, rnnlm=None):
-        """SegmentStreamingE2E constructor.
-
-        :param E2E e2e: E2E ASR object
-        :param recog_args: arguments for "recognize" method of E2E
-        """
-        self._e2e = e2e
-        self._recog_args = recog_args
-        self._char_list = e2e.char_list
-        self._rnnlm = rnnlm
-
-        self._e2e.eval()
-
-        self._blank_idx_in_char_list = -1
-        for idx in range(len(self._char_list)):
-            if self._char_list[idx] == self._e2e.blank:
-                self._blank_idx_in_char_list = idx
-                break
-
-        self._subsampling_factor = np.prod(e2e.subsample)
-        self._activates = 0
-        self._blank_dur = 0
-
-        self._previous_input = []
-        self._previous_encoder_recurrent_state = None
-        self._encoder_states = []
-        self._ctc_posteriors = []
-
-        assert self._recog_args.ctc_weight > 0.0, \
-            "SegmentStreamingE2E works only with combined CTC and attention decoders."
-        assert "b" not in self._e2e.etype, \
-            "SegmentStreamingE2E works only with uni-directional encoders"
-
-    def accept_input(self, x):
-        """Call this method each time a new batch of input is available."""
-
-        self._previous_input.extend(x)
-        h, ilen = self._e2e.subsample_frames(x)
-
-        # Run encoder and apply greedy search on CTC softmax output
-        h, _, self._previous_encoder_recurrent_state = self._e2e.enc(
-            h.unsqueeze(0),
-            ilen,
-            self._previous_encoder_recurrent_state
-        )
-        z = self._e2e.ctc.argmax(h).squeeze(0)
-
-        if self._activates == 0 and z[0] != self._blank_idx_in_char_list:
-            self._activates = 1
-
-            # Rerun encoder with zero state at onset of detection
-            tail_len = self._subsampling_factor * (self._recog_args.onset_margin + 1)
-            h, ilen = self._e2e.subsample_frames(
-                np.reshape(self._previous_input[-tail_len:], [-1, len(self._previous_input[0])]))
-            h, _, self._previous_encoder_recurrent_state = self._e2e.enc(
-                h.unsqueeze(0), ilen, None)
-
-        hyp = None
-        if self._activates == 1:
-            self._encoder_states.extend(h.squeeze(0))
-            self._ctc_posteriors.extend(self._e2e.ctc.log_softmax(h).squeeze(0))
-
-            if z[0] == self._blank_idx_in_char_list:
-                self._blank_dur += 1
-            else:
-                self._blank_dur = 0
-
-            if self._blank_dur >= self._recog_args.min_blank_dur:
-                seg_len = len(self._encoder_states) - self._blank_dur + self._recog_args.offset_margin
-                if seg_len > 0:
-                    # Run decoder with a detected segment
-                    h = torch.cat(self._encoder_states[:seg_len], dim=0).view(
-                        -1, self._encoder_states[0].size(0))
-                    lpz = torch.cat(self._ctc_posteriors[:seg_len], dim=0).view(
-                        -1, self._ctc_posteriors[0].size(0))
-                    hyp = self._e2e.dec.recognize_beam(
-                        h, lpz, self._recog_args, self._char_list, self._rnnlm)
-
-                    self._activates = 0
-                    self._blank_dur = 0
-
-                    tail_len = self._subsampling_factor * self._recog_args.onset_margin
-                    self._previous_input = self._previous_input[-tail_len:]
-                    self._encoder_states = []
-                    self._ctc_posteriors = []
-
-        return hyp
